Pull some methods into new RummagerUrlHelper
We want to break these methods down a bit further and add helpers for
other formats including policies and CMA cases. Moving the methods into
a separate file helps to group them and reduce noise in the
`DocumentHelper` method.
