+++ features/support/document_helper.rb
@@ -55,27 +55,6 @@ module DocumentHelper
     )
   end
 
-  def default_search_params
-    {
-      "count" => "1000",
-      "fields" => mosw_search_fields.join(","),
-      "filter_document_type" => "mosw_report",
-    }
-  end
-
-  def mosw_search_fields
-    %w(
-      title
-      link
-      description
-      public_timestamp
-      walk_type
-      place_of_origin
-      date_of_introduction
-      creator
-    )
-  end
-
   def rummager_all_documents_url
     rummager_url(
       default_search_params.merge(
@@ -110,10 +89,6 @@ module DocumentHelper
       "filter_policies" => ["benefits-reform"],
       "order" => "-public_timestamp",
     )
-  end
-
-  def rummager_url(params)
-    "#{Plek.current.find('search')}/unified_search.json?#{params.to_query}"
   end
 
   def keyword_search_results
